Move quill modules and settings from components to quill service
