+++ frontend/src/app/admin/quill/quill.service.ts
@@ -1,16 +1,55 @@
 import { Injectable } from '@angular/core';
 import { Http, Headers, RequestOptions} from '@angular/http';
 import { MatSnackBar } from '@angular/material';
+import * as Quill from 'quill';
 
 @Injectable({
   providedIn: 'root'
 })
 export class QuillService {
+  quillEditor: any;
 
   IMGUR_CLIENT_ID = 'e52c61fa40312d4';
   IMGUR_API_URL = 'https://api.imgur.com/3/image';
 
-  constructor(private http: Http, private snackBar: MatSnackBar) {  }
+  constructor(private http: Http, private snackBar: MatSnackBar) {
+
+    let Link = Quill.import('formats/link');
+    Link.sanitize = (url) => {
+      const regexp = new RegExp('^(http://|https://)');
+      return regexp.test(url) ? url : 'http://' + url;
+    };
+    Quill.register(Link, true);
+
+    let Embed = Quill.import('blots/embed');
+    class Hr extends Embed {
+      static blotName = 'hr';
+      static tagName = 'hr';
+    }
+    Quill.register(Hr);
+
+   }
+
+  setQuillEditor(editor) {
+    this.quillEditor = editor;
+  }
+
+  getQuillModules() {
+    return { toolbar: {
+      container: [
+        ['bold', 'italic', 'underline'],
+        [{ 'header': 1}, { 'header': 2}],
+        [{ 'list': 'ordered'}, { 'list': 'bullet'}],
+        [{ 'indent': '-1'}, { 'indent': '+1' }], 
+        [{ 'align': [] }],
+        ['link', 'image', 'code']
+      ],
+      handlers: {
+        'image': () => { this.uploadImage(this.quillEditor) },
+        'code' : () => { this.insertHr(this.quillEditor) }
+      }
+    }}
+  }
 
   getRequestOptions(): RequestOptions {
     const headers = new Headers();
